requirements.txt
django==3.0
flake8==6.1.0
Flake8-pyproject==1.2.3
coverage==7.2.7
pytest-django==3.9.0
pytest-cov==4.1
mixer==7.2.2
